Change star repo route token source
Probably due to an unknown CORS issue, since starring a repo is a PUT request,
it does not receive the Authorization header.

Therefore, the token source needs to change from the Authorization header to
a query parameter.

Though not as secure as passing the token in an Authorization header, passing
the token via query param is ok since a typical user cannot copy the URL (user stars
a repo on button click), and also there is protection granted by an SSL cert that is
provided by Google.